+++ spec/osm/api_spec.rb
@@ -419,7 +419,7 @@ describe "API" do
         'token' => @api_config[:api_token],
         'userid' => 'user',
         'secret' => 'secret',
-        'eveningid' => nil, 'sectionid' => nil, 'meetingdate' => '2000-01-02', 'starttime' => nil,
+        'eveningid' => 1, 'sectionid' => 2, 'meetingdate' => '2000-01-02', 'starttime' => nil,
         'endtime' => nil, 'title' => 'Unnamed meeting', 'notesforparents' =>'', 'prenotes' => '',
         'postnotes' => '', 'games' => '', 'leaders' => '', 'activity' => '[]',
       }
@@ -427,7 +427,7 @@ describe "API" do
       api.stub(:get_terms) { [] }
       HTTParty.should_receive(:post).with(url, {:body => post_data}) { DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"result":0}'}) }
 
-      evening = Osm::Evening.new({:meeting_date => Date.new(2000, 01, 02)})
+      evening = Osm::Evening.new(:id=>1, :section_id=>2, :meeting_date=>Date.new(2000, 01, 02))
       api.update_evening(evening).should be_true
     end
 
@@ -438,7 +438,7 @@ describe "API" do
         'token' => @api_config[:api_token],
         'userid' => 'user',
         'secret' => 'secret',
-        'eveningid' => nil, 'sectionid' => nil, 'meetingdate' => '2000-01-02', 'starttime' => nil,
+        'eveningid' => 1, 'sectionid' => 2, 'meetingdate' => '2000-01-02', 'starttime' => nil,
         'endtime' => nil, 'title' => 'Unnamed meeting', 'notesforparents' =>'', 'prenotes' => '',
         'postnotes' => '', 'games' => '', 'leaders' => '', 'activity' => '[]',
       }
@@ -446,10 +446,15 @@ describe "API" do
       api.stub(:get_terms) { [] }
       HTTParty.should_receive(:post).with(url, {:body => post_data}) { DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"result":1}'}) }
 
-      evening = Osm::Evening.new({:meeting_date => Date.new(2000, 01, 02)})
+      evening = Osm::Evening.new(:id=>1, :section_id=>2, :meeting_date=>Date.new(2000, 01, 02))
       api.update_evening(evening).should be_false
     end
-  end
+
+    it "Update an evening (invalid evening)" do
+      api = Osm::Api.new('user', 'secret')
+      evening = Osm::Evening.new
+      expect{ api.update_evening(evening) }.to raise_error(Osm::ArgumentIsInvalid)
+    end  end
 
 
   describe "Options Hash" do